requirements.txt
alembic==1.16.4 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
astroid==3.3.10 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.30.0 ; python_version >= "3.12" and python_version < "4.0"
bandit==1.8.6 ; python_version >= "3.12" and python_version < "4.0"
black==25.1.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.7.9 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
click==8.2.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
cryptography==45.0.5 ; python_version >= "3.12" and python_version < "4.0"
dill==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
ecdsa==0.19.1 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
environs==14.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-cors==0.0.6 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.14 ; python_version >= "3.12" and python_version < "4.0"
flake8==7.3.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.2.3 ; python_version >= "3.12" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
isort==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.10 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
marshmallow==4.0.0 ; python_version >= "3.12" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
mypy==1.16.1 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.12" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
pbr==6.1.1 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.3.8 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycodestyle==2.14.0 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pycryptodome==3.23.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.12" and python_version < "4.0"
pydantic[email]==2.11.7 ; python_version >= "3.12" and python_version < "4.0"
pyflakes==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
pylint==3.3.7 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.4.1 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
python-jose[cryptography]==3.5.0 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
redis==6.2.0 ; python_version >= "3.12" and python_version < "4.0"
rich==14.0.0 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.12" and python_version < "4"
setuptools==80.9.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.41 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy[asyncio]==2.0.41 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.12" and python_version < "4.0"
stevedore==5.4.1 ; python_version >= "3.12" and python_version < "4.0"
tomlkit==0.13.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.34.3 ; python_version >= "3.12" and python_version < "4.0"
